Use package resource, remote_file
Previously a Powershell exec was used to install packages. This commit
converts that task to a package-based resource. It allows cleaner,
easier to read code.

Switch from download_file to remote_file. This allows for better
subscribe/notify of the dotnet defined type as remote_file does not have
any intermediate file fragment resources which may be used or changed
even if the file to download is not changed. The remote_file resource is
also platform agnostic, unlike download_file, more meaning it may be
accessible and familar to a wider range of potential users/contributors.