Add logic to prevent incompatible versions
Some versions of .NET are in-place upgrades of others. Installation of
.NET 4.5, for example, will replace the 4.0 package. In order to
disallow Puppet from continuously trying to install .NET 4.0 in the
event dotnet resources for both 4.0 and 4.5 have been added to the
catalog, create a package=absent resource for each conflicting version.
This will cause the conflict to be caught when a catalog is compiled for
the node.